requirements.txt
opencv-python-headless
camelot-py
streamlit
altair<5